requirements.txt
playwright==1.49.0
python-dotenv==1.0.1
pytest==8.3.4